Added read-in routines for Tokyo Teck spinning disk data
Contained in routine read_TT_TiffStacks.m

All files of a single embryo have to be contained within one folder.
Then the routine is initialized to the folder and maps out the data
structure. Further calls are then used to interact with this structure
and request specific images from color, time point, and z section.